feat: wrap up amount field
