+++ clients/desktop/package.json.md5
@@ -1,1 +1,1 @@
-2437c2127f2c7181198aba2b792a0704
+6b3b26e954e801cfec688caec9e75b1a
